+++ src/main/java/com/b1n_ry/yigd/compat/InvModCompat.java
@@ -17,9 +17,13 @@ public interface InvModCompat<T> {
         FabricLoader loader = FabricLoader.getInstance();
         YigdConfig.CompatConfig compatConfig = YigdConfig.getConfig().compatConfig;
 
+        boolean accessoriesPresent = compatConfig.enableAccessoriesCompat && loader.isModLoaded("accessories");
+
         boolean trinketsPresent = compatConfig.enableTrinketsCompat && loader.isModLoaded("trinkets");
 
-        if (trinketsPresent)
+        if (accessoriesPresent)
+            invCompatMods.add(new AccessoriesCompat());
+        if (trinketsPresent && !loader.isModLoaded("tclayer"))
             invCompatMods.add(new TrinketsCompat());
 /*        if (compatConfig.enableInventorioCompat && loader.isModLoaded("inventorio"))
             invCompatMods.add(new InventorioCompat());*/