+++ src/main/java/com/b1n_ry/yigd/config/YigdConfig.java
@@ -274,22 +274,26 @@ public class YigdConfig implements ConfigData {
         @ConfigEntry.Gui.EnumHandler(option = ConfigEntry.Gui.EnumHandler.EnumDisplayOption.BUTTON)
         public DropRule standardDropRuleInClaim = DropRule.PUT_IN_GRAVE;
 
-        public boolean enableInventorioCompat = true;
+        public boolean enableAccessoriesCompat = true;
         @Comment("While PUT_IN_GRAVE, other drop rules will be prioritized")
         @ConfigEntry.Gui.EnumHandler(option = ConfigEntry.Gui.EnumHandler.EnumDisplayOption.BUTTON)
-        public DropRule defaultInventorioDropRule = DropRule.PUT_IN_GRAVE;
+        public DropRule defaultAccessoriesDropRule = DropRule.PUT_IN_GRAVE;
+//        public boolean enableInventorioCompat = true;
+//        @Comment("While PUT_IN_GRAVE, other drop rules will be prioritized")
+//        @ConfigEntry.Gui.EnumHandler(option = ConfigEntry.Gui.EnumHandler.EnumDisplayOption.BUTTON)
+//        public DropRule defaultInventorioDropRule = DropRule.PUT_IN_GRAVE;
 //        public boolean enableLevelzCompat = true;
 //        @Comment("While PUT_IN_GRAVE, other drop rules will be prioritized")
 //        @ConfigEntry.Gui.EnumHandler(option = ConfigEntry.Gui.EnumHandler.EnumDisplayOption.BUTTON)
 //        public DropRule defaultLevelzDropRule = DropRule.PUT_IN_GRAVE;
-        public boolean enableNumismaticOverhaulCompat = true;
-        @Comment("While PUT_IN_GRAVE, other drop rules will be prioritized")
-        @ConfigEntry.Gui.EnumHandler(option = ConfigEntry.Gui.EnumHandler.EnumDisplayOption.BUTTON)
-        public DropRule defaultNumismaticDropRule = DropRule.PUT_IN_GRAVE;
-        public boolean enableOriginsInventoryCompat = true;
-        @Comment("While PUT_IN_GRAVE, other drop rules will be prioritized")
-        @ConfigEntry.Gui.EnumHandler(option = ConfigEntry.Gui.EnumHandler.EnumDisplayOption.BUTTON)
-        public DropRule defaultOriginsDropRule = DropRule.PUT_IN_GRAVE;
+//        public boolean enableNumismaticOverhaulCompat = true;
+//        @Comment("While PUT_IN_GRAVE, other drop rules will be prioritized")
+//        @ConfigEntry.Gui.EnumHandler(option = ConfigEntry.Gui.EnumHandler.EnumDisplayOption.BUTTON)
+//        public DropRule defaultNumismaticDropRule = DropRule.PUT_IN_GRAVE;
+//        public boolean enableOriginsInventoryCompat = true;
+//        @Comment("While PUT_IN_GRAVE, other drop rules will be prioritized")
+//        @ConfigEntry.Gui.EnumHandler(option = ConfigEntry.Gui.EnumHandler.EnumDisplayOption.BUTTON)
+//        public DropRule defaultOriginsDropRule = DropRule.PUT_IN_GRAVE;
         public boolean enableTravelersBackpackCompat = true;
         @Comment("While PUT_IN_GRAVE, other drop rules will be prioritized")
         @ConfigEntry.Gui.EnumHandler(option = ConfigEntry.Gui.EnumHandler.EnumDisplayOption.BUTTON)